pyproject.toml
[tool.poetry]
name = "fluigi"
version = "0.1.0"
description = "Place and route engine"
authors = ["Radhakrishna Sanka <rkrishnasanka@gmail.com>"]
license = "BSD-Clause-3"

[tool.poetry.dependencies]
python = "^3.8"
antlr4-python3-runtime = "^4.8"
networkx = "^2.5"
argparse = "^1.4.0"
hypernetx = "^0.3.7"
pydot = "^1.4.1"
matplotlib = "^3.3.2"
pyfiglet = "^0.8.post1"
pygraphviz = "^1.6"
JPype1-py3 = "^0.5.5"
requests = "^2.24.0"
pycairo = "^1.20.0"
JPype1 = "^1.1.2"
pymint = {path = "/home/krishna/CIDAR/pymint"}

[tool.poetry.dev-dependencies]
mypy = "^0.790"
black = "^20.8b1"
flake8 = "^3.8.4"
rope = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fluigi = "fluigi.cmdline:main"
